fix: keep type on flat item (#736)
Though the spec [recommends dropping
it](https://github.com/stac-utils/stac-geoparquet/blob/a59ba7638117fc12c12784dbbe7dbd8c2d583b88/spec/stac-geoparquet-spec.md?plain=1#L23),
I'd like to keep it for now:

- There's already at least one other probably-identical string field in
the file (`stac_version`) so if we're trying to optimize on size we
could play with metadata more to cut more columns
- Naïve geoparquet readers are more likely to be able to reconstruct
valid STAC items if all the required fields are present ... without
`type`, you really have to know what you're doing
- In our case, when we search **stac-geoparquet** we don't assume that
we're returning valid STAC items, because the `fields` extension can be
used to strip away required fields. Because of that, a direct conversion
from **stac-geoparquet** to JSON will not "rehydrate" the `type` field.
We _could_ add logic to `search` to do this "rehydration", but it seemed
simpler and not super-high-consequence to just keep the column

cc @hrodmn